fix: use client 오류 수정
